fix: quite a few windows bug fixes

+++ include/webview.hpp
@@ -70,7 +70,7 @@ namespace RenWeb {
                 webview_impl->set_title(title);
             }
             void set_size(int64_t width, int64_t height) override {
-                webview_impl->set_size(width, height, WEBVIEW_HINT_NONE);
+                webview_impl->set_size(static_cast<int>(width), static_cast<int>(height), WEBVIEW_HINT_NONE);
             }
             void set_html(const std::string& html) override {
                 webview_impl->set_html(html);
